Backend/requirements.txt
Django==4.2.0
djangorestframework==3.15
djangorestframework-simplejwt==5.3.1
setuptools==75.6.0
psycopg2-binary==2.9.10
sqlparse==0.5.2
asgiref==3.8.1
django-extensions==3.2.3
django-cors-headers==4.6.0